💥 Add a distinction between exceptions and errors
There are several distinctions between exceptions and errors.

- An exception is a known but unexpected outcome of a risky operation.
  On the other hand, an error is a programming mistake.
- Exceptions need to be handled. Errors need to be fixed.
- An exception becomes an error if it's not handled. An error becomes an
  exception when its caught at an error boundary.

In order to distinguish between exceptions and errors, thereby making
the library more robust, I renamed the `CustomError` class to
`Exception`. I also changed the way we construct instances of
`Exception`. Furthermore, I changed the input of `unsafeExtract`. It now
expects an `Exception` instead of any error. Finally, I removed the
`UnsafeExtractError` class from the list of exports because it's an
error. Hence, the user shouldn't be able to catch it. If the user wants
to catch the error thrown by `unsafeExtract` then they should provide a
custom instance of `Exception`.