✨ Add `Expression` type and the `evaluate` function
The `Expression` type is a reification of an unevaluated expression,
i.e. a thunk, or a non-function value. The value must be a non-function
because otherwise the `Expression` type wouldn't be a disjoint union.
The evaluate function return the value of a given expression. It will
evaluate and return the result of a thunk. On the other hand, it will
directly return non-function values.

The `Expression` type provides a very convenient way to create type-safe
functions whose inputs are optionally lazy. The caller of the function
can decide whether they want to provide a strict input or a lazy input.
They can also do so for each argument separately. It's also useful for
API designers because they don't have to create separate functions for
strict and lazy inputs. It also makes it easier to test the functions.