💥 Make the `Option` and `Result` constructors lazy
I made the `Some`, `Success`, and `Failure` constructors lazy by
updating them to accept expressions. This is a breaking change because
the constructors can't accept non-thunk function values anymore. Hence,
I also created `Some.of`, `Success.of`, and `Failure.of` static methods
which can be used for the old behavior. This is great because we no
longer need to use the `new` keyword and we can pass these static
methods around like regular functions. I also updated the tests and even
added a few new tests.