✅ Check whether the error is a `TypeError` on fail
Converting an arbitrary value to a string can result in a `TypeError` if
the value doesn't have a callable `toString` or if `toString` returns an
object, and if it doesn't have a callable `valueOf` or if `valueOf`
returns an object. Hence, we need to add a test for this error.