💥 Replace instance functions with instance methods
The core philosophy of this library is to embrace methods instead of
functions. Having separate instances for setoids, partial orders, and
total orders went against this philosophy. In this PR, I rectified that.
The tradeoff is that we need to box primitives before we can use their
instance methods. This is similar to newtypes in Haskell. This is an
acceptable tradeoff because it gives developers the freedom to choose
or implement their own boxed types.