✨ Add `fromGenerator` function and `effect` method
The `fromGenerator` function and the `effect` method can be used to
write monadic code à la do notation in Haskell or `async…await` syntax
in JavaScript. The `fromGenerator` function converts a generator
function into an `Option` and the `effect` method converts an `Option`
into an effectful generator so that it can be used with `yield*` for
side effectful code. I also added an `effectMap` function which maps the
result of the effectful generator after yielding. This function is just
for convenience.