💥 Replace `unsafeExtract` and rename `safeExtract`
The `unsafeExtract` function is unwieldy. We can replace it with the
more generic `mapExtractSome` function which takes a `getDefaultValue`
thunk as an input. The `getDefaultValue` function can then explicitly
throw an error if it wants to, recovering the original behavior of the
`unsafeExtract` function. Since we replace `unsafeExtract` with the
`mapExtractSome` function, I also renamed `safeExtract` to `extractSome`
for consistency.