💥 Add `transpose`, `toResult` & `toOption` methods
The `Option#transpose` method converts an `Option` to a `Result`.
Similarly, the `Result#transpose` method converts a `Result` back to an
`Option`. However, there's more than one way to convert an `Option` to a
`Result` and back. Hence, I added the `transpose`, `transposeOkay`, and
`transposeFail` methods for both `Option` and `Result` and also their
`map` variants.

Similarly, we have the `Option#toResult` and `Result#toOption` methods.
Again, there's more than one way to convert an `Option` to a `Result`
and back. Hence, I added the `toResultOkay` and `toResultFail` methods,
and their `toOption` counterparts.