✨ Add the `and` and `or` methods to `ResultTrait`
A `Result` is an applicative functor in two different ways. The `and`
method is the applicative functor of `Result` with respect to `Okay.of`.
The `or` method is the applicative functor of `Result` with respect to
`Fail.of`. In addition, I added the `andThen`, `andWhen`, `orElse`, and
`orErst` methods because they are commonly used. I also fixed a typo in
the `option.test.ts` file.