💥 Remove the `transpose` and `transposeMap` methods
The `transpose` and `transposeMap` methods aren't lawful because the
`Result` type is not a `Biapplicative`. The only reason we can define it
is because the `Option` container has at most one element. Hence, we
never need to combine multiple elements. Also, for `None` we arbitrarily
chose to return an `Okay`. We could have also chosen to return a `Fail`.
Hence, the `transpose` and `transposeMap` functions aren't satisfactory
or even particularly useful.