💥 Remove union type from `and`, `or`, and variants
When the return type is a union, it's easier to use the API. However,
it's also a common source of bugs. Removing the union type doesn't make
the API any less powerful. You can still have a return union type if you
want to. You just have to be explicit about it. This prevents the user
from making accidental mistakes because they need to be explicit about
the return type. It also makes the type signature easier to read and
understand, which makes the library more consistent and useable overall.