🐛 Fix the return type of the `commute` method
When the input of `commute` is `Result<A, B>` then the output should be
`Result<B, A>`. However, I accidentally set the output as `Result<B,
B>`. Unfortunately, TypeScript did not catch this mistake because it
doesn't properly check overloaded function types. Hence, we need to be
extra careful when writing overloaded functions.