✨ Add `fromGenerator`, `effect`, and `effectMap`
The `Result` type is a deterministic monad. Hence, we can use generators
as syntactic sugar to work with the `Result` type. To that end, I added
the `Okay.fromGenerator` static function, and `effect` and `effectMap`
methods. The only catch is that the `Fail` value will always be
`unknown`, which is fair because any error will be automatically
converted to a `Fail` value and we don't know what error can be thrown.