💥 Swap the order of generics of the `Result` type
In `Result<E, A>`, the `E` was the type of the value in `Fail` whereas
the `A` was the type of the value in `Okay`. This was getting confusing
because in all the methods the `Okay` value always comes first. Hence,
there's cognitive dissonance because we need to switch the order for
some methods. Thus, I swapped the order of generics. In `Result<A, E>`,
the `A` is the type of the value in `Okay` whereas the `E` is the type
of the value in `Fail`. This leads to consistency between the order of
the generics in the type and the order of arguments to methods of the
`Result` type, leading to less cognitive dissonance. In addition, this
is also the order of generics in Rust. This is a major breaking change.
Not only did I change the order of the generics, but I also renamed
`associateLeft` to `associateRight` and vice versa.