💥 Rename the `Pair.of` method to `Pair.from`
The `of` static method should have two inputs because it denotes the
identity element of a `Biapplicative` functor. However, `Pair.of` has
only one input which it duplicates to create a `Pair`. Hence, I renamed
`Pair.of` to `Pair.from` so that I can create a separate `Pair.of`
method which has two inputs.