💥 Rename the `[Symbol.iterator]` method to `values`
The `[Symbol.iterator]` method is implicitly called whenever we use
`for…of` or the spread operator. However, it doesn't play nicely with
generics. Since we're using generics for type inference, we lose type
information whenever we use `for…of` or the spread operator. Hence, I
renamed the `[Symbol.iterator]` method to `values`. Since we now have to
explicitly call `.values()` whenever we want to use `for…of` or the
spread operator, type inference works correctly. In addition, explicit
is better than implicit. It makes the code more readable.