✨ Add `uncurry2`, `uncurry3`, & `uncurry4` functions
The `uncurry2`, `uncurry3`, and `uncurry4` functions are meant to be
used in conjunction with the applicative interface of data types. For
example, you can write `u.and(v).and(w).and(x).map(uncurry4(f))`. This
is equivalent to `f <$> u <*> v <*> w <*> x` in Haskell. We only provide
`uncurry` functions up to 4 parameters. If the programmer needs more
parameters, then they are expected to define their own `uncurry`
functions. The expectation is that `uncurry2`, `uncurry3`, and
`uncurry4` should be able to handle 99% of use cases.