♻️ Remove `implements` from `Option` and `Result`
The `implements` clause was used to ensure that the class implements all
the methods of the specified interfaces. However, there's no good reason
to do that because if the class doesn't implement the interfaces then
the tests would fail. In addition, the `implements` clause is incorrect.
The implementation isn't for `Option<never>` and `Result<never, never>`.
Hence, it's more accurate to not specify the `implements` clause at all.