💥 Remove `Some.fromGenerator` and `NoneException`
I removed `Some.fromGenerator` because it throws a `NoneException` when
the generate yields `None`. This is not ideal because `Option` does not
implement `MonadError`. Hence, we shouldn't be able to catch the `None`
value. If developers want to use generators then they should convert the
`Option` into a `Result` and use `Okay.fromGenerator` instead. In doing
so, they have to explicitly provide an error. Hence, it's better.